LPS-157478 Add tooltipTitle attribute to the node for the tooltip
