+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page-structure/components/StructureTreeNode.js
@@ -20,7 +20,6 @@ import React, {useEffect, useRef, useState} from 'react';
 
 import {addMappingFields} from '../../../../../app/actions/index';
 import {fromControlsId} from '../../../../../app/components/layout-data-items/Collection';
-import {EDITABLE_TYPES} from '../../../../../app/config/constants/editableTypes';
 import {REQUIRED_FIELD_DATA} from '../../../../../app/config/constants/formModalData';
 import {ITEM_ACTIVATION_ORIGINS} from '../../../../../app/config/constants/itemActivationOrigins';
 import {ITEM_TYPES} from '../../../../../app/config/constants/itemTypes';
@@ -66,15 +65,6 @@ import useHasRequiredChild from '../../../../../app/utils/useHasRequiredChild';
 import useControlledState from '../../../../../core/hooks/useControlledState';
 
 const HOVER_EXPAND_DELAY = 1000;
-
-const EDITABLE_LABEL = {
-	[EDITABLE_TYPES.backgroundImage]: Liferay.Language.get('background-image'),
-	[EDITABLE_TYPES.html]: Liferay.Language.get('html'),
-	[EDITABLE_TYPES.image]: Liferay.Language.get('image'),
-	[EDITABLE_TYPES.link]: Liferay.Language.get('link'),
-	[EDITABLE_TYPES['rich-text']]: Liferay.Language.get('rich-text'),
-	[EDITABLE_TYPES.text]: Liferay.Language.get('text'),
-};
 
 const loadCollectionFields = (
 	dispatch,
@@ -282,8 +272,6 @@ function StructureTreeNodeContent({
 		};
 	}, [isOverTarget, node]);
 
-	const isEditable = node.itemType === ITEM_TYPES.editable;
-
 	return (
 		<div
 			aria-disabled={node.isMasterItem || !node.activable}
@@ -322,16 +310,9 @@ function StructureTreeNodeContent({
 				aria-label={Liferay.Util.sub(Liferay.Language.get('select-x'), [
 					node.name,
 				])}
-				className={classNames(
-					'page-editor__page-structure__tree-node__mask',
-					{
-						'lfr-portal-tooltip': isEditable,
-					}
-				)}
-				data-title={
-					isEditable ? EDITABLE_LABEL[node.editableType] : null
-				}
-				data-tooltip-align={isEditable ? 'left' : null}
+				className="lfr-portal-tooltip page-editor__page-structure__tree-node__mask"
+				data-title={node.tooltipTitle}
+				data-tooltip-align="left"
 				onClick={() => {
 					const itemId = getFirstControlsId({
 						item: node,
